+++ ChaosMod/Mp3Manager.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <string>
+
 namespace Mp3Manager
 {
 	void PlayChaosSoundFile(const std::string& szSoundFile);
